Moved Bad Apple to a new file
